# API-Spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Employee Task Management REST API",
    "version": "1.0.0",
    "description": "REST API documentation for Employee Task Management application with JWT security."
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "summary": "Register a new user",
        "tags": [
          "auth-controller"
        ],
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                },
                "examples": {
                  "registerAdminSuccess": {
                    "summary": "Register Admin  - Success",
                    "value": {
                      "message": "User registered successfully",
                      "userId": 9,
                      "email": "whohania12@gmail.com",
                      "role": "ADMIN"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "registerEmptyPassword": {
                    "summary": "Register - Empty Password",
                    "value": {
                      "message": "Password cannot be empty",
                      "status": 400,
                      "timestamp": "2026-08-06T02:11:51.404855"
                    }
                  },
                  "registerInvalidEmailFormat": {
                    "summary": "Register - Invalid Email Format",
                    "value": {
                      "message": "Invalid email format",
                      "status": 400,
                      "timestamp": "2026-08-06T02:09:02.502435"
                    }
                  },
                  "registerEmptyName": {
                    "summary": "Register - Empty Name",
                    "value": {
                      "message": "Name cannot be empty",
                      "status": 400,
                      "timestamp": "2026-08-06T02:06:27.710201"
                    }
                  },
                  "registerEmptyEmail": {
                    "summary": "Register - Empty Email",
                    "value": {
                      "message": "Email cannot be empty",
                      "status": 400,
                      "timestamp": "2026-08-06T02:07:55.050257"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "registerInvalidToken": {
                    "summary": "Register - Invalid Token",
                    "value": {
                      "message": "Authentication failed",
                      "status": 401,
                      "timestamp": "2026-08-06T02:18:29.576493"
                    }
                  },
                  "registerAdminSuccess": {
                    "summary": "Register Admin  - Success",
                    "value": {
                      "message": "Authentication failed",
                      "status": 401,
                      "timestamp": "2026-08-06T02:16:09.959423"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Requires ADMIN role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "registerAccessDenied": {
                    "summary": "Register - Access Denied",
                    "value": {
                      "message": "Access denied. You do not have permission to access this resource.",
                      "status": 403,
                      "timestamp": "2026-08-06T02:20:36.26191"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict - User already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "registerAdminSuccess": {
                    "summary": "Register Admin  - Success",
                    "value": {
                      "message": "User Already Exist",
                      "status": 409,
                      "timestamp": "2026-08-06T02:04:23.548743"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Authenticate user and get JWT token",
        "tags": [
          "auth-controller"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                },
                "examples": {
                  "loginAdminSuccess": {
                    "summary": "Login - Admin Success",
                    "value": {
                      "token": "eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJhZG1pbkBnbWFpbC5jb20iLCJyb2xlIjoiQURNSU4iLCJpYXQiOjE3ODU5NjA1MDIsImV4cCI6MTc4NjA0NjkwMn0.dN-nkznZJqfwt--e7wlEsduFM0B7Sw0jj7fcTkam6Gdt1nxHFgReLN93edFthGSXtiihPs8wdizZKPM2YC-hqg",
                      "role": "ADMIN",
                      "userId": 1,
                      "userName": "admin@gmail.com"
                    }
                  },
                  "loginEmployeeSuccess": {
                    "summary": "Login - Employee Success",
                    "value": {
                      "token": "eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJhbWFzaEBnbWFpbC5jb20iLCJyb2xlIjoiRU1QTE9ZRUUiLCJpYXQiOjE3ODU5NjA3MjgsImV4cCI6MTc4NjA0NzEyOH0.GBnSzx5xXsQ3Km0hLM3-bFLindMBrKwxX70gpE9eXK8Y3SDwuo2XKwcZX1BDYwo1isvnvmzHU6qQ3U2RencydA",
                      "role": "EMPLOYEE",
                      "userId": 3,
                      "userName": "amash@gmail.com"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "loginEmptyEmail": {
                    "summary": "Login - Empty Email",
                    "value": {
                      "message": "Email cannot be empty",
                      "status": 400,
                      "timestamp": "2026-08-06T01:48:09.386852"
                    }
                  },
                  "loginEmptyPassword": {
                    "summary": "Login - Empty Password",
                    "value": {
                      "message": "Password cannot be empty",
                      "status": 400,
                      "timestamp": "2026-08-06T01:52:49.748169"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid email or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "loginInvalidPasswordOrEmail": {
                    "summary": "Login - Invalid Password Or Email",
                    "value": {
                      "message": "Invalid email or password",
                      "status": 401,
                      "timestamp": "2026-08-06T01:45:00.422965"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "maxLength": 2147483647,
            "minLength": 8
          }
        },
        "required": [
          "email",
          "name",
          "password"
        ]
      },
      "RegisterResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "userName": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  }
}