# backend/src/main/java/com/zohaib/taskmanager/dto/LoginRequest.java
package com.zohaib.taskmanager.dto;

import jakarta.validation.constraints.NotBlank;

public class LoginRequest {
	@NotBlank(message = "Email cannot be empty")
	String email;
	
	@NotBlank(message = "Password cannot be empty")
    String password;
    
	public String getEmail() {
		return email;
	}
	public void setEmail(String email) {
		this.email = email;
	}
	public String getPassword() {
		return password;
	}
	public void setPassword(String password) {
		this.password = password;
	}
    
    
}