# backend/src/main/java/com/zohaib/taskmanager/dto/LoginResponse.java
package com.zohaib.taskmanager.dto;

public class LoginResponse {
	String token;
    String role;
    Long userId;
    String userName;
    
	public String getToken() {
		return token;
	}
	public void setToken(String token) {
		this.token = token;
	}
	public String getRole() {
		return role;
	}
	public void setRole(String role) {
		this.role = role;
	}
	public Long getUserId() {
		return userId;
	}
	public void setUserId(Long userId) {
		this.userId = userId;
	}
	public String getUserName() {
		return userName;
	}
	public void setUserName(String userName) {
		this.userName = userName;
	}
    
    
}