# backend/src/main/java/com/zohaib/taskmanager/dto/RegisterResponse.java
package com.zohaib.taskmanager.dto;

public class RegisterResponse {

	private String message;
    private Long userId;
    private String email;
    private String role;
    
	public String getMessage() {
		return message;
	}
	public void setMessage(String message) {
		this.message = message;
	}
	public Long getUserId() {
		return userId;
	}
	public void setUserId(Long userId) {
		this.userId = userId;
	}
	public String getEmail() {
		return email;
	}
	public void setEmail(String email) {
		this.email = email;
	}
	public String getRole() {
		return role;
	}
	public void setRole(String role) {
		this.role = role;
	}
    
    
}